bitsy-py/requirements/requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
backports.cached-property==1.0.2
base58==2.1.1
bitarray==1.2.2
blake3==0.3.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
cytoolz==0.11.2
eth-abi==3.0.0
eth-account==0.6.1
eth-hash[pycryptodome]==0.3.3
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.1.0
eth-utils==2.0.0
fastapi==0.78.0
frozenlist==1.3.0
graphql-core==3.2.1
h11==0.12.0
hexbytes==0.2.2
httpcore==0.13.7
httpx==1.0.0b0
hvac==0.11.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.1
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pygments==2.12.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pyyaml==6.0
requests==2.28.1
rfc3986[idna2008]==1.5.0
rlp==3.0.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.116.2
toolz==0.11.2
typing-extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
varint==1.0.2
web3==6.0.0b3
websockets==10.3
yarl==1.7.2
